make external source (application/helper) to load automatically without specified name-spaces in config ini

+++ application/helpers/Processor/Mergerates.php
@@ -120,9 +120,9 @@ class Processor_Mergerates extends Billrun_Processor_Base_Separator {
 				$value = array(
 					'unit' => $unit,
 					'rate' => array(
-						"to" => 2147483647,
-						"price" => $row['tinf_sampPrice0'],
-						"interval" => $row['tinf_sampDelayInSec0'],
+						"to" => (int) 2147483647,
+						"price" => (int) $row['tinf_sampPrice0'],
+						"interval" => (int) $row['tinf_sampDelayInSec0'],
 					),
 				);
 				$entity->set("rates.".$rateKey, $value);
